requirements.txt
acres
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
async-timeout
attrs
blinker
cachetools
certifi
charset-normalizer
ci-info
click
configobj
configparser
dataclasses-json
distro
etelemetry
exceptiongroup
filelock
frozenlist
gitdb
GitPython
groq
h11
httpcore
httplib2
httpx
httpx-sse
idna
importlib_resources
isodate
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
langchain-groq
langchain-ollama
langchain-text-splitters
langsmith
looseversion
lxml
markdown-it-py
MarkupSafe
marshmallow
mdurl
multidict
mypy-extensions
narwhals
networkx
nibabel
nipype
numpy
ollama
orjson
packaging
pandas
pathlib
pillow
propcache
protobuf
prov
puremagic
pyarrow
pydantic
pydantic-settings
pydantic_core
pydeck
pydot
Pygments
PyMuPDF
pyparsing
python-dateutil
python-dotenv
pytz
pyxnat
PyYAML
rdflib
referencing
requests
requests-toolbelt
rich
rpds-py
scipy
simplejson
six
smmap
sniffio
SQLAlchemy
streamlit
tenacity
toml
tornado
traits
typing-inspect
typing_extensions
tzdata
urllib3
yarl
zipp
zstandard
google-generativeai